refactor: semaphore use usize (#32)

+++ mea/src/rwlock/mod.rs
@@ -83,7 +83,7 @@ pub use write_guard::RwLockWriteGuard;
 /// See the [module level documentation](self) for more.
 pub struct RwLock<T: ?Sized> {
     /// Maximum number of concurrent readers.
-    max_readers: u32,
+    max_readers: usize,
     /// Semaphore to coordinate read and write access to T
     s: Semaphore,
     /// The inner data.
@@ -128,7 +128,7 @@ impl<T> RwLock<T> {
     /// ```
     pub fn new(t: T) -> RwLock<T> {
         // large enough while not touch the edge
-        let default_max_readers = u32::MAX >> 1;
+        let default_max_readers = usize::MAX >> 1;
         RwLock::with_max_readers(t, default_max_readers)
     }
 
@@ -144,7 +144,7 @@ impl<T> RwLock<T> {
     ///
     /// let rwlock = RwLock::with_max_readers(5, 1024);
     /// ```
-    pub fn with_max_readers(t: T, max_readers: u32) -> RwLock<T> {
+    pub fn with_max_readers(t: T, max_readers: usize) -> RwLock<T> {
         let s = Semaphore::new(max_readers);
         let c = UnsafeCell::new(t);
         RwLock { max_readers, c, s }
