+++ mea/src/rwlock/write_guard.rs
@@ -90,7 +90,7 @@ impl<T: ?Sized> RwLock<T> {
 /// This structure is created by the [`RwLock::write`] method.
 #[must_use = "if unused the RwLock will immediately unlock"]
 pub struct RwLockWriteGuard<'a, T: ?Sized> {
-    pub(super) permits_acquired: u32,
+    pub(super) permits_acquired: usize,
     pub(super) lock: &'a RwLock<T>,
 }
